Windows build fix.

+++ src/player/AreaNode.cpp
@@ -180,7 +180,7 @@ float AreaNode::getAngle() const
 
 void AreaNode::setAngle(float angle)
 {
-    m_Angle = fmod(angle, 2*M_PI);
+    m_Angle = fmod(angle, 2*(float)M_PI);
     m_bTransformChanged = true;
 }
